Consolidated media wiki api calls in a single place
