+++ app/src/main/java/fr/free/nrw/commons/contributions/ContributionsSyncAdapter.java
@@ -19,8 +19,8 @@ import java.util.ArrayList;
 import java.util.Date;
 
 import fr.free.nrw.commons.CommonsApplication;
-import fr.free.nrw.commons.MWApi;
 import fr.free.nrw.commons.Utils;
+import fr.free.nrw.commons.mwapi.MediaWikiApi;
 import timber.log.Timber;
 
 public class ContributionsSyncAdapter extends AbstractThreadedSyncAdapter {
@@ -61,7 +61,7 @@ public class ContributionsSyncAdapter extends AbstractThreadedSyncAdapter {
     public void onPerformSync(Account account, Bundle bundle, String s, ContentProviderClient contentProviderClient, SyncResult syncResult) {
         // This code is fraught with possibilities of race conditions, but lalalalala I can't hear you!
         String user = account.name;
-        MWApi api = CommonsApplication.getInstance().getMWApi();
+        MediaWikiApi api = CommonsApplication.getInstance().getMWApi();
         SharedPreferences prefs = this.getContext().getSharedPreferences("prefs", Context.MODE_PRIVATE);
         String lastModified = prefs.getString("lastSyncTimestamp", "");
         Date curTime = new Date();
@@ -71,19 +71,7 @@ public class ContributionsSyncAdapter extends AbstractThreadedSyncAdapter {
         while(!done) {
 
             try {
-                MWApi.RequestBuilder builder = api.action("query")
-                        .param("list", "logevents")
-                        .param("letype", "upload")
-                        .param("leprop", "title|timestamp|ids")
-                        .param("leuser", user)
-                        .param("lelimit", getLimit());
-                if(!TextUtils.isEmpty(lastModified)) {
-                    builder.param("leend", lastModified);
-                }
-                if(!TextUtils.isEmpty(queryContinue)) {
-                    builder.param("lestart", queryContinue);
-                }
-                result = builder.get();
+                result = api.logEvents(user, lastModified, queryContinue, getLimit());
             } catch (IOException e) {
                 // There isn't really much we can do, eh?
                 // FIXME: Perhaps add EventLogging?
